- fixing null pointer in edit form wizard - getting edit form to show in the action list in the content forms directory as well
git-svn-id: https://svn.alfresco.com/repos/alfresco-enterprise/alfresco/HEAD/root@4685 c4b6b30b-aa2e-2d43-bbcb-ca4b014f7261

+++ source/java/org/alfresco/web/bean/wcm/EditFormWizard.java
@@ -152,9 +152,12 @@ public class EditFormWizard
                                       fileInfo.getNodeRef());
       }
 
-      for (RenderingEngineTemplateData retd : this.removedRenderingEngineTemplates)
+      if (this.removedRenderingEngineTemplates != null)
       {
-         this.nodeService.removeChild(formNodeRef, retd.getNodeRef());
+         for (RenderingEngineTemplateData retd : this.removedRenderingEngineTemplates)
+         {
+            this.nodeService.removeChild(formNodeRef, retd.getNodeRef());
+         }
       }
 
       for (RenderingEngineTemplateData retd : this.renderingEngineTemplates)
